Fix(Store): creates store.js and fixes circular import
