+++ client/store/pool.js
@@ -1,6 +1,6 @@
 import axios from 'axios'
 import history from '../history'
-import {LOADING} from '.'
+import {LOADING} from './store'
 const asyncHandler = require('express-async-handler')
 /**
  * ACTION TYPES
